+++ test-compute-kernels/CHANGELOG.md
@@ -9,6 +9,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 ### Changed
+### Deprecated
+### Removed
+### Fixed
+### Security
+
+## [1.7.0] - 2023-08-03
+### Added
+### Changed
 - Updating the conan dependencies versions to released components
 ### Deprecated
 ### Removed